requirements.txt
bcrypt
beautifulsoup4
bs4
cryptography
email-validator
fastapi
httpx
matplotlib
numpy
opencv-python
pandas
passlib
pillow
psycopg2-binary
pydantic
pydantic_core
PyJWT
requests
SQLAlchemy
tqdm
uvicorn
python-dotenv
fastapi-jinja
sentence-transformers
keras-ocr   
tensorflow
speechrecognition